Move PositionMapper back into compilercommon
We need to use it directly in the generator
